test(bench): measure production selective parsing
