+++ bench/scriptData.bench.ts
@@ -1,86 +1,22 @@
 import { bench, describe } from 'vitest';
 import { parseScriptData } from '../src/core/scriptData.js';
-import type { SpecMap } from '../src/core/spec.js';
-import { appSpecs } from '../src/features/app/specs.js';
+import { createApp } from '../src/features/app/app.js';
+import { appScriptDataSelection } from '../src/features/app/specs.js';
 import { APP_FIXTURES, loadAppFixture } from './fixtures.js';
 import type { AppFixtureName } from './fixtures.js';
 
-const SCRIPT_BLOCK_REGEX = />AF_initDataCallback[\s\S]*?<\/script/g;
-const BLOCK_KEY_REGEX = /(ds:.*?)'/;
-const BLOCK_PAYLOAD_REGEX = /data:([\s\S]*?), sideChannel: {}}\);<\//;
-
-const specs: SpecMap = appSpecs;
-
-function referencedDsKeys(): Set<string> {
-  const keys = new Set<string>();
-  for (const spec of Object.values(specs)) {
-    for (const path of spec.paths) {
-      const head = path[0];
-      if (typeof head === 'string' && head.startsWith('ds:')) {
-        keys.add(head);
-      }
-    }
-  }
-  return keys;
-}
-
-function parseBlocksOnly(html: string): Record<string, unknown> {
-  const blocks: Record<string, unknown> = {};
-  const matches = html.match(SCRIPT_BLOCK_REGEX);
-  if (matches === null) {
-    return blocks;
-  }
-  for (const block of matches) {
-    const key = BLOCK_KEY_REGEX.exec(block)?.[1];
-    const payload = BLOCK_PAYLOAD_REGEX.exec(block)?.[1];
-    if (key === undefined || payload === undefined) {
-      continue;
-    }
-    try {
-      blocks[key] = JSON.parse(payload);
-    } catch {
-      continue;
-    }
-  }
-  return blocks;
-}
-
-function parseReferencedBlocks(
-  html: string,
-  referenced: ReadonlySet<string>,
-): Record<string, unknown> {
-  const blocks: Record<string, unknown> = {};
-  const matches = html.match(SCRIPT_BLOCK_REGEX);
-  if (matches === null) {
-    return blocks;
-  }
-  for (const block of matches) {
-    const key = BLOCK_KEY_REGEX.exec(block)?.[1];
-    if (key === undefined || !referenced.has(key)) {
-      continue;
-    }
-    const payload = BLOCK_PAYLOAD_REGEX.exec(block)?.[1];
-    if (payload === undefined) {
-      continue;
-    }
-    try {
-      blocks[key] = JSON.parse(payload);
-    } catch {
-      continue;
-    }
-  }
-  return blocks;
-}
+const offlineApp = (html: string) => createApp(() => ({ request: () => Promise.resolve(html) }));
 
 const sink = { total: 0 };
 
 for (const name of Object.keys(APP_FIXTURES) as AppFixtureName[]) {
   const html = loadAppFixture(name);
-  const referenced = referencedDsKeys();
+  const appId = APP_FIXTURES[name];
+  const app = offlineApp(html);
 
   describe(name, () => {
     bench(
-      'parseScriptData',
+      'parseScriptData all blocks',
       () => {
         sink.total += Object.keys(parseScriptData(html).blocks).length;
       },
@@ -88,17 +24,18 @@ for (const name of Object.keys(APP_FIXTURES) as AppFixtureName[]) {
     );
 
     bench(
-      'parse blocks only',
+      'parseScriptData selected blocks',
       () => {
-        sink.total += Object.keys(parseBlocksOnly(html)).length;
+        sink.total += Object.keys(parseScriptData(html, appScriptDataSelection).blocks).length;
       },
       { time: 1000 },
     );
 
     bench(
-      'parse referenced blocks only',
-      () => {
-        sink.total += Object.keys(parseReferencedBlocks(html, referenced)).length;
+      'app() offline',
+      async () => {
+        const result = await app({ appId });
+        sink.total += result.title.length;
       },
       { time: 1000 },
     );
